Implemented movement as a group
One archon is elected the “master archon” at the beginning of the game.
It starts moving in a random direction and randomly changes direction
when meets the boundary of the map. It broadcasts its moving direction
(almost) every turn in a 30 square radius. Soldiers and other archons
getting this message and which are not attacking or building other
soldiers now move to be a few squares ahead of the master archon, where
ahead refers to the direction the master archon broadcasts (in addition
to its position, which is part of every signal in Battlecode 2016).