Enemy search + soldier attack
Implemented a locateEnemy() method for the master archon. This method
returns if the archon has the enemy location. If not, it asks scouts
for the enemy location. If the master archon gets no answers after a
number of turns, it asks normal archons to build more scouts, and waits
another while before asking scouts for locations again.

Changed soldier attack routine to let it stray from pack a little bit
and be more aggressive.