SPRINT TOURNAMENT SUBMISSION
This is the player that was uploaded for the Sprint Tournament one week
after the specs for Battlecode 2016 were released.

Changes in this commit:

* Scouts will now broadcast the enemy location anytime after they
receive a enemy position request, rather than only on the turn they
receive it.

* If they are not too far away from the target location indicated by an
Archon, Soldiers will attack the weakest enemy in range, or walk
towards the weakest enemy they can see. My implementation for this
casts senseHostileRobots() twice, one with the sensor range and one
with the attack range, because for some reason I could not code a
reliable method that extracts the enemies within attack range from
those the Soldier can see. This is embarrassing. To be implemented in
the future: make soldiers attack the opponent first and zombies second.

* Normal archons will walk to parts if they can sense it before
attempting to build new robots. This uses a findClosest() method added
to Movement.java.