Deleted unused code
Deleted an implementation of a path finding algorithm of breadth first
type in the Movement.java file.

Deleted the findSweetSpot() method from the Soldier.java file. Its
description, taken from comments in the code:

We search in the disk of radius rc.getType().attackRadiusSquared
for a place that is as far away from enemy as possible. Given the
choice of two map locations that are the same distance away from the
enemy, we elect the one that is closest to us. The map locations of
nearby robots are not eligible to be the sweet spot.