Solder in danger signal: call and respose
Made enemies, allies, and enemiesInRange arrays static fields of the
Soldier class.

Implemented a simple callForHelp() method that simply broadcasts a
signal (without message) in a small radius. Any signal without a
message is interpreted as a distress signal from a soldier.

This method callForHelp() is called when a soldier senses more enemies
than allies nearby.

Added a static field called soldiersInDanger, implemented as a queue of
map locations.

Created a static field called incomingSignals which is updated at the
beginning of the execution of the repeat() method.

incomingSignals is fed to a updateSoldiersInDanger( Signals[] ) method
that adds an element to the queue of soldiers in danger for each signal
without a message in incomingSignals.

updateSoldiersInDanger removes a soldier from the list of those in
danger if it the Robot Controller can sense it, but no nearby enemies.

Finally, changed the repeat() method. If there are nearby enemies, it
calls fight(). Otherwise it moves to the location of the soldier in
danger in front of the queue soldiersInDanger.