Built more of the soldier AI
I aim to implement the following AI for a soldier. At each turn the
Soldier will do the first of the following actions whose contains are
met.

1) If there are enemies nearby, fight them.

2) If there are Archon distress signals in the queue called
“archonsInDanger”, attack move to the signal’s location.

3) If there are Soldier distress signals in the queue called
“soldiersInDanger”, attack-move to the signal’s location.

4) If there are any known zombie den locations, move towards them.

5) Do nothing.

—

The archon mechanic has not yet been implemented, but the others have
to a satisfactory degree. Details of the current state of the
implementation:

* At the beginning of each turn, soldiers update static variables that
keep track nearby robots. They sense nearby robots once and process
update the variables associated to the different kinds of robots using
the updateNearbyRobots() method.

* The next step is to update incoming signals. This means calling
rc.emtpySignalQueue() and converting the result to a LlinkedList of
Signals. The latter data structure results in code that is easier to
read when doing loops and allows for easy element removal doing loops
using Iterators.

* The soldier class now has a static boolean field called
sentASignalThisTurn. This is due to the way in which soldiers signals
are implemented: a single one in a turn represents a call for help, two
indicate that the soldier can sense a zombie den, and three indicate
that the soldier can either sense spare parts or neutral robots. Thus a
robot can send a single message per turn.

* There is an updateTasks() method that goes through the field
incomingSignals and counts signals from individual soldiers to update
the following fields: archonsInDanger, soldiersInDanger, zombieDens,
parts, and neutralRobots. This method’s implementation it as of yet
incomplete and in particular does not handle most archon signals.

* updateTasks() also calls checkForCompletedTasks(), which for example
removes a soldier from the list of those in danger if the robot
controller is in attacking range of the soldier’s location when it sent
a distress signal, but cannot see enemies.

* There is a also a check for dens method, which well, does that: if
there are any nearby and there are no zombie den locations reported in
the area, add the robot controller’s location to the reported areas and
broadcast two signals.