Started rewriting archon code
So far the archon moves away from enemies when it can see any, and
tries to build a soldier when left alone.

The archon code is being written by making changes to a copy of the
soldier code. Both classes have many classes in common. Some the archon
will use, other not. The latter will be deleted from the Archon class
in the future, whereas the former will eventually compose utility
classes.