Restructuring the code
I grouped together the readings (of robots, signals and parts) into a
class called well, Readings. This class also contains methods used for
updating and processing the information that is sensed.

This commit has the new class. The next one will update the Soldier
class to make use of the new Readings class.