SEEDING TOURNAMENT SUBMISSION
Sort of. The actual submission was somewhere between this and the last
commit, but closest to this one. This iteration — and the submission, I
think — represents a downgrade from the previous commit, as most test
matches between the two versions of the code will show. So it’s
unfortunate that this iteration got submitted and not the last one. The
reason it got submitted is that these were last-minute “improvements”
and were submitted after only very preliminary tests. Further testing
after the deadline closed revealed “bugs” and the fact that this bot
will lose in mosts fights agains the previous one.

The changes that were made were the following. I extend the previously
developed Goals class to handle parts and neutral robots in addition to
zombie dens.

* The implementation of parts collection by archons was initially
shortened by using this class. It was then modified to accommodate the
other goals that archons will now pursue, that of activating neutral
robots.

The gravest problem introduced was the following. Archons will now try
to move towards goals, and if they are behind enemies, they will freeze
between the goals and soldiers. Also substantially, archons are
constantly generating exceptions. This is possibly due to them going
over the byte code limit of 20k per turn (which I am at this point sure
they are doing).

COMMENTS ON THE LAST COMMIT

In the soldier class there are static fields for nearby enemies,
enemies in range, allies, zombie dens, parts as well as incoming
signals. These were bundled into a class called Readings, together with
methods for managing them.

The fields storing archons and soldiers in danger, known zombie dens,
and known parts and neutral robots locations, as well as the methods
that manage them, were bundled in the Goals class.