Small fixes and improvements
Archons can now ask for soldiers to clear the way for them to pass. The
radius of this broadcast is larger if the archon has negative patience.

If soldiers stop shooting a den to fight something else, they will
remember to go back and finish up the den for some time.

Fixed the reading of enemies in range.

Fixed randomDirection method. It is possibly using excessive byte codes
as is (but works).

The range of call for help increases if you have more allies near you.