Communication
The TeamGoals class is being overhauled and currently does not work. It
is being changed to work with a request-reply system. Thus if a soldier
knows of no zombie den positions, it will ask its colleagues for one by
sending three signals out. Similarly, an Archon will ask its colleagues
for the position of resources (parts and neutral robots) by means of a
message signal. Only upon receiving such a request that a robot will
broadcast the location of a zombie den or resources. Robots which
cannot send message signals will reply with two signals (which carry
their current location), only if they can sense a zombie den or
resources during the turn they receive the request. Robots which can
send message signals will broadcast the locations of the last zombie
dens or resource locations they have seen, depending on whether the
request was for dens (without message) or for resources (with message).

The reason for this request-reply system is that it will result in
fewer messages being broadcast. While broadcasting messages every turn
is cheap for any given robot, processing messages that all other
friendly robots are needlessly broadcasting is expensive.

In making this change I also learned something about hash tables, so
let me interweave what I learned with the changes in the code.

The current iteration of the code uses a hash table to implement a
histogram in the method that extract goals from signals. Thus the
table’s keys are the signals received from robots with different id’s
and the table’s values record the number of times a given id occurs.

Retrieval of a given key is essentially a search problem. Search is
much easier if the set being searched is ordered. Thus keys are
associated to integers (hash values), and as the set of keys is being
built, it is ordered using these hash values. This makes retrieval of
the value of a given key much easier.

To implement the histogram, I had to change what it means for two
Signals to be considered equal. This was achieved by creating a new
class which contains a single Signal as a field and which has its own
implementation of (overrides) the equals( Object other ) method. This
new class also overrides hashCode(), which now simply returns
signal.getID().