Communication II
Created a simple Timer class to be used throughout the code.

Changed the names of the method canSeeNewDen to goalsFromDenReadings,
and similarly for parts, neutrals and signals.

Implemented methods for soldiers to ask others for den locations and
for archons to ask other units for resource locations (i.e., locations
for parts and neutral robots).

Added comments which explain the code and comments that separate
different parts of the code.

Made it so that if one robot can tell another already replied a
question it received, it won’t reply itself.

Created a method for adding map locations to the goal lists (dens,
parts and neutrals).

Added technology for transmitting map locations via message signals.