Spiral scouting
Called the earlier scouting method “random scouting” and introduced a
new one, spiral scouting. The old method consisted choosing a direction
randomly, moving along it for a fixed amount of time or until the end
of the map is reached. Spiral scouting consists of moving following the
pattern:

R U LL DD RRR UUU LLLL DDDD RRRRR UUUUU …

Here R stands for ‘right’, U for ‘up, and so on, and movement along R
means in fact 7 steps or so, which is about the Scout’s sensor range.