Added vipers!
Instead of focusing on the weakest enemy within range, vipers shoot as
many different opponents as possible with the pupose of infecting them.
They keep track of each enemy’s infection by means of a hash table with
keys the ID’s of those enemy robots that are infected, and with values
independent timers that count the infection time remaining.

Slight change: scouts now also tell soldiers to make way for them.